test(io): move config related behavior tests to test_config.py
